frontend-app-api: migrate createExtensionInstance tests
